refactor: move logic to useProduct hook
